### 2019-10-24 #1836 @dfm
@kthyng: sure thing!

@GreenScarz: Please feel free to make suggestions for reviewers! You can find a list of potential reviewers [here](https://bit.ly/joss-reviewers) Otherwise, I'll start looking shortly.
