### 2019-10-24 #1836 @GreenScarz
Hi @dfm and @kthyng , Thanks for taking the time to chat :) all of the analytical tools included therein are tools which I've used to publish materials research before; the biggest change has just been repackaging everything for general use. So really, anyone who has a knack for being hypercritical about documentation and then trying to break things will suffice as a reviewer. In the test folder is example data which they can use when poking around at the library and reviewing all the software.
