### 2019-10-25 #1836 @GreenScarz
Hey does anyone here know if changing your GitHub username will negatively effect this review/publication process? Obviously I made this account before I realized that it would be used for professional functions haha, probably shouldn't be tethered to my gamertag.
