### 2019-10-25 #1836 @1mikegrn
so new username is @1mikegrn , the @GreenScarz links are discontinued and I've updated the docs in the libRL library to contain the new pip installation instructions.
