### 2019-10-26 #1836 @lucydot
Hi @dfm I'm currently in the middle of another JOSS review, I'd rather not commit to two.
