### 2019-11-03 #1836 @1mikegrn
No worries @dfm, I understand review process takes time. Looked into your comment a bit, glad you pointed that out because I was previously oblivious to OS dependencies for compiled cython files - implemented a patch which should clear everything up and mitigate OS dependencies from cython.
