### 2019-11-06 #1868 @whedon
[REVIEW]: libRL: A Python library for the characterization of microwave absorption
**Submitting author:** @1mikegrn (<a href="http://orcid.org/0000-0002-2525-1274">Michael Green</a>)
**Repository:** <a href="https://github.com/1mikegrn/libRL-package" target ="_blank">https://github.com/1mikegrn/libRL-package</a>
**Version:** v1.1.0
