### 2019-11-07 1mikegrn/libRL#1 @kjappelbaum
Paths do not work on all OS
Hi, as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1868) I'll open some small 'bug' issues and later one with more comments.

The first one is that the way you join paths, e.g, in the `examples.py` does not work on all operating systems. It would be better if you use an OS agnostic way (`os.path.join`).
