### 2019-11-07 1mikegrn/libRL-package#2 @kjappelbaum
LICENSE file not plain text and not recognized by pip 
Hi, again as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1868).

Your license is not a plaintext file (`LICENSE`), but a markdown. this is likely the reason also GitHub does not show it in the upper right corner.

Also, pip does not show the license, in my case the output of `pip show` is

```
Name: libRL
Version: 1.0.3
Summary: Library of functions used for characterizing Microwave Absorption
Home-page: https://github.com/1mikegrn/libRL
Author: Michael Green
Author-email: 1mikegrn@gmail.com
License: UNKNOWN
Location: /Users/kevinmaikjablonka/anaconda3/envs/joss_librl/lib/python3.7/site-packages
Requires: numpy, pandas, pathos, xlrd, cython, scipy, matplotlib
Required-by:
```
