### 2019-11-07 1mikegrn/libRL-package#3 @kjappelbaum
PEP8 / linting and autoformatting 
Hi, a minor point I noticed looking at the code (for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1868)) is that it does follow PEP8 in all places.

Even though it is not an explicit criterion it makes working with the code easier if one sticks to PEP8.

This can be done relatively easily with some auto formatting upon each commit using e.g. the pre-commit library which you can then use to run linters like prospector and code formatters like yapf or black. You could also use this to automatically run tests (locally) upon each commit.

In particular, functions should be lowercase according to PEP8 and also type checks might be better written in duck type flavor in `try/except` cases or if `isinstance` rather than with `type` (see e.g. [stackoverflow](https://stackoverflow.com/questions/1549801/what-are-the-differences-between-type-and-isinstance))

Below, I'm just copy-pasting the prospector output so you can give a look if you think that there are points that are reasonable to fix.  Most of them are really minor.

```
Messages
========

libRL/__init__.py
  Line: 44
    pylint: wrong-import-position / Import "import pyximport" should be placed at the top of the module
  Line: 46
    pep8: E275 / missing whitespace after keyword (col 18)
    pylint: import-self / Module import itself
    pylint: wrong-import-position / Import "from libRL import refactoring, quick_graphs, cpfuncs" should be placed at the top of the module
  Line: 52
    pylint: redefined-builtin / Redefining built-in 'abs'
    pylint: wrong-import-position / Import "from numpy import arange, zeros, abs, array, float64, errstate, pi, sqrt" should be placed at the top of the module
  Line: 57
    pylint: wrong-import-position / Import "from pandas import DataFrame" should be placed at the top of the module
  Line: 58
    pylint: wrong-import-position / Import "from pathos.multiprocessing import ProcessPool as Pool" should be placed at the top of the module
  Line: 236
    pylint: literal-comparison / Comparison to literal (col 11)
  Line: 291
    pylint: too-many-locals / Too many local variables (19/15)
  Line: 387
    pylint: unidiomatic-typecheck / Using type() instead of isinstance() for a typecheck. (col 7)
  Line: 486
    pylint: too-many-locals / Too many local variables (17/15)
  Line: 632
    pep8: E306 / expected 1 blank line before a nested definition, found 0 (col 5)

libRL/quick_graphs.py
  Line: 29
    pylint: unused-import / Unused mpl_toolkits.mplot3d.axis3d imported as axis3d
  Line: 64
    pylint: protected-access / Access to a protected member _axinfo of a client class (col 8)
  Line: 120
    pep8: E722 / do not use bare 'except' (col 5)
  Line: 127
    pep8: E722 / do not use bare 'except' (col 5)
  Line: 173
    pep8: E722 / do not use bare 'except' (col 5)
  Line: 180
    pep8: E722 / do not use bare 'except' (col 5)
  Line: 185
    pylint: trailing-newlines / Trailing newlines

libRL/refactoring.py
  Line: 38
    pylint: redefined-builtin / Redefining built-in 'abs'
  Line: 99
    pylint: bare-except / No exception type(s) specified (col 20)
  Line: 164
    pylint: literal-comparison / Comparison to literal (col 30)
  Line: 224
    pylint: literal-comparison / Comparison to literal (col 9)
  Line: 238
    pylint: literal-comparison / Comparison to literal (col 9)
  Line: 276
    pylint: no-else-raise / Unnecessary "elif" after "raise" (col 4)
  Line: 281
    pylint: unidiomatic-typecheck / Using type() instead of isinstance() for a typecheck. (col 9)
  Line: 289
    pep8: E722 / do not use bare 'except' (col 9)
  Line: 317
    pylint: unidiomatic-typecheck / Using type() instead of isinstance() for a typecheck. (col 7)
  Line: 323
    pep8: E722 / do not use bare 'except' (col 9)
  Line: 337
    pylint: no-else-raise / Unnecessary "else" after "raise" (col 4)

Check Information
=================
         Started: 2019-11-07 09:11:28.638680
        Finished: 2019-11-07 09:11:39.781930
      Time Taken: 11.14 seconds
       Formatter: grouped
        Profiles: default, strictness_medium, strictness_high, strictness_veryhigh, no_doc_warnings, no_test_warnings, no_member_warnings
      Strictness: medium
  Libraries Used:
       Tools Run: dodgy, mccabe, pep8, profile-validator, pyflakes, pylint
  Messages Found: 31
```

and the output from just running pylint

```
************* Module libRL
__init__.py:211:0: C0330: Wrong continued indentation (add 41 spaces).
            (e1f(f) - cmath.sqrt(-1) * e2f(f)))) * (cmath.tanh(j *
            ^                                        | (bad-continuation)
__init__.py:212:0: C0330: Wrong continued indentation (add 51 spaces).
            (2 * cmath.pi * (f * 10**9) * (d * 0.001) / 299792458) *
            ^                                                  | (bad-continuation)
__init__.py:213:0: C0330: Wrong continued indentation (add 51 spaces).
            cmath.sqrt((mu1f(f) - j * mu2f(f)) * (e1f(f) - j *
            ^                                                  | (bad-continuation)
__init__.py:214:0: C0330: Wrong continued indentation (add 38 spaces).
            e2f(f)))))) - 1) / ((1 * (cmath.sqrt((mu1f(f) - j * mu2f(f)) /
            ^                                     | (bad-continuation)
__init__.py:215:0: C0330: Wrong continued indentation (add 37 spaces).
            (e1f(f) - j * e2f(f)))) * (cmath.tanh(j * (2 *
            ^                                    | (bad-continuation)
__init__.py:216:0: C0330: Wrong continued indentation (add 43 spaces).
            cmath.pi * (f * 10**9) * (d * 0.001) / 299792458) * cmath.sqrt(
            ^                                          | (bad-continuation)
__init__.py:217:0: C0330: Wrong hanging indentation (add 4 spaces).
            (mu1f(f) - j * mu2f(f)) * (e1f(f) - j *
            ^   | (bad-continuation)
__init__.py:218:0: C0330: Wrong continued indentation (add 27 spaces).
            e2f(f)))))) + 1)))))
            ^                          | (bad-continuation)
__init__.py:249:0: C0330: Wrong hanging indentation before block.
            ) is True:
    |   |   ^ (bad-continuation)
__init__.py:271:0: C0330: Wrong hanging indentation (remove 24 spaces).
                                        i * gridInt:(i + 1) * gridInt, 0:3
                |                       ^ (bad-continuation)
__init__.py:272:0: C0330: Wrong hanging indentation.
                                        ]
            |   |                       ^ (bad-continuation)
__init__.py:635:0: C0330: Wrong continued indentation (add 5 spaces).
            (e1f(f) - cmath.sqrt(-1) * e2f(f))).real)) * (
            ^    | (bad-continuation)
__init__.py:656:0: C0330: Wrong hanging indentation before block.
            ) is True:
    |   |   ^ (bad-continuation)
__init__.py:659:19: C0326: No space allowed around keyword argument assignment
            d_vals = d_set,
                   ^ (bad-whitespace)
__init__.py:660:19: C0326: No space allowed around keyword argument assignment
            m_vals = m_set,
                   ^ (bad-whitespace)
__init__.py:52:0: W0622: Redefining built-in 'abs' (redefined-builtin)
__init__.py:1:0: C0103: Module name "libRL" doesn't conform to snake_case naming style (invalid-name)
__init__.py:42:0: C0103: Constant name "here" doesn't conform to UPPER_CASE naming style (invalid-name)
__init__.py:44:0: C0413: Import "import pyximport" should be placed at the top of the module (wrong-import-position)
__init__.py:44:18: C0321: More than one statement on a single line (multiple-statements)
__init__.py:46:0: C0413: Import "from libRL import refactoring, quick_graphs, cpfuncs" should be placed at the top of the module (wrong-import-position)
__init__.py:46:0: W0406: Module import itself (import-self)
__init__.py:52:0: C0413: Import "from numpy import arange, zeros, abs, array, float64, errstate, pi, sqrt" should be placed at the top of the module (wrong-import-position)
__init__.py:57:0: C0413: Import "from pandas import DataFrame" should be placed at the top of the module (wrong-import-position)
__init__.py:58:0: C0413: Import "from pathos.multiprocessing import ProcessPool as Pool" should be placed at the top of the module (wrong-import-position)
__init__.py:61:0: C0103: Function name "RL" doesn't conform to snake_case naming style (invalid-name)
__init__.py:61:0: C0103: Argument name "Mcalc" doesn't conform to snake_case naming style (invalid-name)
__init__.py:205:4: C0103: Function name "G" doesn't conform to snake_case naming style (invalid-name)
__init__.py:206:8: C0103: Variable name "f" doesn't conform to snake_case naming style (invalid-name)
__init__.py:207:8: C0103: Variable name "d" doesn't conform to snake_case naming style (invalid-name)
__init__.py:210:8: C0103: Variable name "y" doesn't conform to snake_case naming style (invalid-name)
__init__.py:236:11: R0123: Comparison to literal (literal-comparison)
__init__.py:259:8: C0103: Variable name "gridInt" doesn't conform to snake_case naming style (invalid-name)
__init__.py:264:8: C0103: Variable name "MCres" doesn't conform to snake_case naming style (invalid-name)
__init__.py:291:0: C0103: Function name "CARL" doesn't conform to snake_case naming style (invalid-name)
__init__.py:291:0: C0103: Argument name "Mcalc" doesn't conform to snake_case naming style (invalid-name)
__init__.py:291:0: R0914: Too many local variables (19/15) (too-many-locals)
__init__.py:387:7: C0123: Using type() instead of isinstance() for a typecheck. (unidiomatic-typecheck)
__init__.py:394:4: C0103: Variable name "c" doesn't conform to snake_case naming style (invalid-name)
__init__.py:395:4: C0103: Variable name "GHz" doesn't conform to snake_case naming style (invalid-name)
__init__.py:396:4: C0103: Variable name "Z0" doesn't conform to snake_case naming style (invalid-name)
__init__.py:397:4: C0103: Variable name "e0" doesn't conform to snake_case naming style (invalid-name)
__init__.py:467:4: C0103: Variable name "Matrix" doesn't conform to snake_case naming style (invalid-name)
__init__.py:486:0: C0103: Function name "BARF" doesn't conform to snake_case naming style (invalid-name)
__init__.py:486:0: C0103: Argument name "Mcalc" doesn't conform to snake_case naming style (invalid-name)
__init__.py:486:0: R0914: Too many local variables (17/15) (too-many-locals)
__init__.py:621:4: C0103: Variable name "PnPGrid" doesn't conform to snake_case naming style (invalid-name)
__init__.py:632:4: C0103: Argument name "f" doesn't conform to snake_case naming style (invalid-name)
__init__.py:632:4: C0103: Argument name "m" doesn't conform to snake_case naming style (invalid-name)
__init__.py:633:8: C0103: Variable name "y" doesn't conform to snake_case naming style (invalid-name)
__init__.py:640:4: C0103: Variable name "mGrid" doesn't conform to snake_case naming style (invalid-name)
__init__.py:643:11: C0103: Variable name "m" doesn't conform to snake_case naming style (invalid-name)
```
