### 2019-11-07 1mikegrn/libRL#1 @kjappelbaum
Somewhat related, I do not know if it is good to hard code the file separator without mentioning it (but there might also be good reasons for doing so). E.g., here https://github.com/1mikegrn/libRL-package/blob/3596a1530dd423e46748e62f14f2f06378099a06/libRL/refactoring.py#L91

I often see files that have the ending `csv` put use `tsv` formatting. Pandas seems to be good in figuring it out automatically.
