### 2019-11-07 1mikegrn/libRL#4 @kjappelbaum
Tests could be more fine-grained and automatized 
Hi, again for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1868).

I saw that you have a test case for the code, which is great! If you want to improve on that you can write unit tests that test each (relevant) function (this makes it easier for new contributors to see where exactly they break your code) and run those tests automatically, e.g. with [pytest](https://pytest.org) on different platforms using e.g. [travis](https://travis-ci.org) or [azure pipelines](https://docs.microsoft.com/en-us/azure/devops/pipelines/ecosystems/python).
