### 2019-11-07 1mikegrn/libRL#1 @kjappelbaum
Also 'Times New Roman' is [not available on e.g. Linux by default](https://github.com/1mikegrn/libRL-package/blob/d7a0b7b8cf35c56c0e797429fe14fc31c941a303/libRL/quick_graphs.py#L32). This might cause minor issues

https://github.com/1mikegrn/libRL-package/blob/d7a0b7b8cf35c56c0e797429fe14fc31c941a303/libRL/quick_graphs.py#L32

But I see that you have "warnings" in this module that it is primarily for you.
