### 2019-11-07 1mikegrn/libRL#5 @kjappelbaum
Contribution guide is missing
Another point for the [review for JOSS](https://github.com/openjournals/joss-reviews/issues/1868): You should probably add a contribution guide. See the [Github documentation](https://help.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors).
