### 2019-11-07 1mikegrn/libRL#3 @kjappelbaum
Also, I'm not sure about the file with this header

https://github.com/1mikegrn/libRL-package/blob/364f07e0cca6fedcde0cad27019314d26fd43c11/libRL/quick_graphs.py#L1-L21

One could consider factoring it out of the main package if you do not want to support it and if it is not ready to use for other the community.

Similar for

> This library is in the alpha stage of development and not available to users outside of direct collaboration. If you wish to use CompGen in your research, please contact Michael at magwwc@mail.umkc.edu to establish a formal collaboration.

in the [documentation](https://1mikegrn.github.io/DocSite/CompGen/)
