### 2019-11-07 1mikegrn/libRL#3 @kjappelbaum
Something that also fits into this issue, as it is also a minor technical point, is that you can likely reduce the number of dependencies by using `concurrent.futures` instead of `pathos` for concurrent processing.
