### 2019-11-07 1mikegrn/libRL#6 @kjappelbaum
Documentation refactoring suggestions
Again for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1868).
I like that you use sphinx and readthedocs for the documentation and have a lot of comments and docstrings.

One idea for improvement of the documentation would be to first give some more introduction and walk the reader through one example (show input, output, reference to theory) and then have some automatically generated API documentation ([sphinx.ext.napoleon](https://sphinxcontrib-napoleon.readthedocs.io/))  better separated from the installation and examples.
