### 2019-11-07 1mikegrn/libRL#7 @kjappelbaum
Example suggestions
Here's one more issue for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1868).

You have an example and it seems to work (besides the path issue). To make it even better you can consider having some more comments on it (e.g., a section in the docs that also shows the output plots one should find in the end) or even a Jupyter notebook with plots and markdown fields that describe what happens (and which can run on [binder](https://mybinder.org)) to show people what the code can do without them having to install the package.
