### 2019-11-07 1mikegrn/libRL#8 @kjappelbaum
JOSS review general comments
Hi, just some general comments on your package, to summarize [my review](https://github.com/openjournals/joss-reviews/issues/1868). I'm not an expert in RAMs  but your package seems to be useful for your community and the software paper does its job in summarizing this need and you also highlight your previous works in it. It also seems to work, besides the minor [path issue](https://github.com/1mikegrn/libRL-package/issues/1).

I think the major points that would make it more user-friendly is to add some verbosity to an examples (see issue https://github.com/1mikegrn/libRL-package/issues/7 or https://github.com/1mikegrn/libRL-package/issues/6). You could even think about adding an installable command line interface as entry point, e.g. using [click](https://palletsprojects.com/p/click/) to make running the analysis a one-liner (at least for the next release). For developer friendliness, you can have a look at some of [output of pylint or prospector](https://github.com/1mikegrn/libRL-package/issues/3) beyond adding the  [contribution guidelines](https://github.com/1mikegrn/libRL-package/issues/5).

There are some issues that I couldn't check (yet) on the review checklist like [automatic tests](https://github.com/1mikegrn/libRL-package/issues/4), [contribution guidelines](https://github.com/1mikegrn/libRL-package/issues/5) and the [licence](https://github.com/1mikegrn/libRL-package/issues/2). But they seem to be relatively easy to fix.
You might also consider adding your package to pypi.
