### 2019-11-07 1mikegrn/libRL#8 @mgdotdev
First off, thank you for taking the time to review this software, I'm looking forward to reading your report and addressing the issues raised accordingly.

I'll comment/address each issue in the thread where it was raised, and then once all the points have been addressed/fixed in the library I'll push the update up to GitHub and comment again here.
