### 2019-11-07 1mikegrn/libRL#8 @mgdotdev
Actually now that I'm thinking about it, I'll probably periodically push instances to Git throughout the editing process so I can easily pull the library down to my Linux VM for cross-platform testing, but I'll be sure to respond here a final time once I've addressed all the comments.
