### 2019-11-07 1mikegrn/libRL#8 @kjappelbaum
just let me know if something is unclear, or I can support you. Many of the comments are just general suggestions for improvement which might make the package a bit nicer.
