### 2019-11-07 1mikegrn/libRL#1 @mgdotdev
> The first one is that the way you join paths, e.g, in the `examples.py` does not work on all operating systems. It would be better if you use an OS agnostic way (`os.path.join`).

- I've gone ahead and edited the example.py file to make use of path.join as the method for concatenating directory strings for that initial test.

> I do not know if it is good to hard code the file separator without mentioning it (but there might also be good reasons for doing so)

- This was an attempt to make explicit to readers that the library expected enforcement of the 'comma' in .csv for when an error was thrown, though now I notice that even with the sep=',' command, if you give it tabbed data, it'll handle the import the same. What I've done to remedy this issue of dealing with .tsv formatted .csv files is to import the file with read_csv, and if the column shape is 1 and a row therein is a string which contains "\t", redo the import with sep = '\t'.

> 'Times New Roman' is [not available on e.g. Linux by default](https://github.com/1mikegrn/libRL-package/blob/d7a0b7b8cf35c56c0e797429fe14fc31c941a303/libRL/quick_graphs.py#L32). This might cause minor issues

- updated the rcParams for cross-compatibility. Still want it to generally work for everyone, just want to avoid formatting complaints; if one person wants the z scale to line up differently than someone else, or have a different color scheme, etc. But ya, rcParams should still work cross-platform for the generic user, so having a font that works for the generic user is a worthwhile patch. Appreciate you pointing that out.
