### 2019-11-07 1mikegrn/libRL#2 @mgdotdev
> Your license is not a plaintext file (`LICENSE`), but a markdown. this is likely the reason also GitHub does not show it in the upper right corner

- Went ahead and updated the license file through GitHub so the site recognizes it as GPL-3.0.

> Also, pip does not show the license

- updated the setup.py file to include the license variable, now visible in `pip show`
