### 2019-11-07 1mikegrn/libRL#6 @kjappelbaum
A further suggestions, that might be useful/make your life easier, is that you could add the docs directly to this repository (as a 'docs' subdirectory) and then [automatically trigger](https://acctestdocs.readthedocs.io/en/latest/webhooks.html) a rebuild (of the API documentation) if you push here.
