### 2019-11-07 1mikegrn/libRL#3 @mgdotdev
So there were a bunch of details herein so I'll try to address them type by type:

>  Line: 44
>     pylint: wrong-import-position / Import "import pyximport" should be placed at the top of the module

- corrected accordingly

> Line: 57
>     pylint: wrong-import-position / Import "from pandas import DataFrame" should be placed at the top of the module

- corrected all the import issues accordingly

> Line: 291
>     pylint: too-many-locals / Too many local variables (19/15)

- ignoring, every variable is necessary

> libRL/quick_graphs.py

- corrected everything therein

> Line: 164
>     pylint: literal-comparison / Comparison to literal (col 30)

the try: except: catchs I'm using herein I'm going to leave for the following reasons:
- on line 103 it's used to bypass the error associated with redefining types in the array and I want to ignore the error that throws.
- in refactoring tuples to numpy arrays it's enforcing input syntax and offering remedy instruction for necessary formatting. The only error I expect here are errors which I'm calling out (and I can't seem to find a better way to do this, will take suggestions though)

> functions should be lowercase

- corrected accordingly, though I'll miss the Avengers reference.

> Also, I'm not sure about the file with this header (quick_graphs.py)

- it's not that it doesn't *work*, the plotting mechanisms work perfectly fine from a technical aspect. The message is simply to deter request for patches which are 'artistic' instead of technical. I've updated the docstring accordingly to better reflect this.

 > Similarly for... in the documentation.

- I structured the DocSite to host docstrings for multiple projects; CompGen is a separate project entirely and the DocSite relevant to this particular library is strictly at '~/DocSite/libRL/'
