### 2019-11-07 1mikegrn/libRL#3 @kjappelbaum
Thanks for giving it a look!

> I structured the DocSite to host docstrings for multiple projects; CompGen is a separate project entirely and the DocSite relevant to this particular library is strictly at '~/DocSite/libRL/'

To reduce confusion I would maybe try to separate then those projects and [have a separate `docs` directory for this repo that is used to build its own readthedocs site](https://github.com/1mikegrn/libRL-package/issues/6).

But I think that you addressed the code style issues (which anyway were only minor points) - so I close this issue.
