### 2019-11-08 1mikegrn/libRL#6 @kjappelbaum
Nice improvement!
