### 2019-11-08 1mikegrn/libRL#5 @mgdotdev
Added a contributing.md file
