### 2019-11-08 1mikegrn/libRL#7 @mgdotdev
I added some comments and references to the DocSite, as well as changed the data source to grab data from the GitHub repository directly so we don't need to worry about path issues. I'm currently looking into hosting jupyter notebooks and binder but atm I think the site and example files should be sufficient.
