### 2019-11-08 1mikegrn/libRL#4 @mgdotdev
I'll look into travis/pytest integration in the future, but for now I added a bit to the libRL_test so to provide full test coverage of the library.
