### 2019-11-08 1mikegrn/libRL#8 @mgdotdev
I think that I've addressed all of the issues presented to an acceptable degree - if you agree, feel free to close the issue raised, or if you find that changes are still necessary, let me know what needs to be further addressed.
