### 2019-11-10 1mikegrn/libRL#1 @kjappelbaum
I feel that there are still some paths that could be handled a bit more elegant, e.g. https://github.com/1mikegrn/libRL/blob/7c9342497b424e9ff5d03ab9fddfbd8816e4404f/libRL/quick_graphs.py#L162
