### 2019-11-10 1mikegrn/libRL#7 @kjappelbaum
Nice improvement! I do not think that Jupyter/Binder are requirements, but they would be nice if you want the library to be widely used.
