### 2019-11-10 1mikegrn/libRL#6 @kjappelbaum
I think the checklist requirements are fulfilled. Still, for the users it might be less confusing if the documentation for separate projects is kept separate. Also, I think using Napoleon would help the sphinx site.
