### 2019-11-10 1mikegrn/libRL#4 @kjappelbaum
Yes, automated testing with coverage report with a badge on the README would be useful. And testing the addition of test would be could (might be nicer to keep them separate to track down what fails. But the project is small so it doesn't matter than much).

But I still do not know if just checking the best test case.
