### 2019-11-11 1mikegrn/libRL#1 @mgdotdev
fixed the image paths
