### 2019-11-11 1mikegrn/libRL#4 @mgdotdev
Okay, was finally able to get all that working haha.

Added travis-CI integration and documentation through sphinx.
