### 2019-11-12 1mikegrn/libRL#8 @kjappelbaum
I think I'm fine now with most points and close this issue.
