### 2019-11-18 #1868 @1mikegrn
Hey @matthewturk, just hoping to get an update on some sort of timetable we can expect for conducting the peer-review. I'm getting ready to submit a set of high-impact research manuscripts and I'd like to be able to reference this manuscript therein for more coverage of the software. Thanks :)
