### 2019-11-18 #1868 @dfm
I think you mean @mdoucet!

@mdoucet: Can you let us know what sort of time frame you need to finish going through the checklist? Thanks!
