### 2019-11-24 1mikegrn/libRL#8 @kjappelbaum
Just a minor point. If you want to keep the acknowledgment in the paper, my name is Kevin M. Jablonka ;)
