### 2019-11-27 #1868 @dfm
@kjappelbaum, @mdoucet: I see that you have completed the checklist - thanks!! Do you have anything else you would like to add or any final recommendations?

@1mikegrn: Looking good! I can't seem to track down that Naito (1969) reference. Is it available online anywhere? Either way, can you now generate a new Zenodo release where the title and author list match this submission. Once you do that, can you post the new DOI and version number to this thread?
