### 2019-11-27 #1868 @dfm
Also: can you all take a last quick look at the paper to make sure that the current version looks good?
