### 2019-11-27 #1868 @1mikegrn
Ya it took me about a month to track it down myself, ended up finding it though a university in Tel Aviv who sent me a pdf of their copy; it's the original paper the mathematics is legible, but the text is in Japanese.

[About the thickness of the ferrite absorption wall - Naito.pdf](https://github.com/openjournals/joss-reviews/files/3897903/About.the.thickness.of.the.ferrite.absorption.wall.-.Naito.pdf)

Also, what exactly needs to go in the Zenodo release? is it the entire package or just the .md file which points to the repository?
