### 2019-11-27 #1868 @dfm
Thanks! The archive should be a snapshot of the full source. The easiest way to do it is following this guide: https://guides.github.com/activities/citable-code/
