### 2019-12-02 #1868 @danielskatz
Once @dfm tells @openjournals/jose-eics that this is ready to publish, one of us (this week, me) will perform the final steps.
