### 2019-12-03 #1868 @dfm
Pinging @openjournals/joss-eics to say that this is ready for the final processing.
